refactor: split _app code to multiple files
PT-1795
