+++ src/domain/app/Center.tsx
@@ -1,0 +1,17 @@
+const Center: React.FC<{ children?: React.ReactNode }> = ({ children }) => {
+  return (
+    <div
+      style={{
+        width: '100vw',
+        height: '100vh',
+        display: 'flex',
+        alignItems: 'center',
+        justifyContent: 'center',
+      }}
+    >
+      {children}
+    </div>
+  );
+};
+
+export default Center;
