+++ src/domain/app/ErrorBoundary.ts
@@ -1,0 +1,25 @@
+import * as Sentry from '@sentry/browser';
+import React, { ErrorInfo } from 'react';
+
+class ErrorBoundary extends React.Component<{ children?: React.ReactNode }> {
+  componentDidCatch(error: Error, errorInfo: ErrorInfo) {
+    if (process.env.NODE_ENV !== 'production') {
+      // eslint-disable-next-line no-console
+      console.log(error);
+    }
+
+    Sentry.withScope((scope) => {
+      scope.setExtra('componentStack', errorInfo.componentStack);
+
+      Sentry.captureException(error);
+    });
+
+    super.componentDidCatch?.(error, errorInfo);
+  }
+
+  render() {
+    return this.props.children;
+  }
+}
+
+export default ErrorBoundary;
